requirements.txt
PyQt5
requests
beautifulsoup4
Pillow
requests-html
selenium
undetected-chromedriver
distutils